refactor: move the network tests into :beer_network and fix the deps they exposed
BeersRemoteSourceTest and TestMockWebService lived in :app, two modules away
from the code they cover. Move them plus their json resources into
:beer_network/src/test, where the remote source they exercise lives.

The move made the module's real dependencies visible. Main never used the
Retrofit converter or the logging interceptor - only the hand-built test stack
does - so both drop to testImplementation, and retrofit2 becomes an explicit
dependency instead of arriving transitively through the converter. androidx.annotation
was unused outright. :app loses the four test dependencies that moved with the
tests, plus kotlinx-serialization-json, which nothing there imports.
