+++ beer_network/src/test/resources/json/fake_typologies_response.json
